chore(audit): fix critical compilation errors - Riverpod providers, missing classes, BuildContext extensions, fromMap methods

+++ lib/services/story_service.dart
@@ -150,7 +150,7 @@ class StoryService {
             (data['expiresAt'] as Timestamp).toDate().isAfter(DateTime.now())) {
           activeStories++;
         }
-        totalViews += data['viewsCount'] ?? 0;
+        totalViews += (data['viewsCount'] ?? 0) as int;
       }
 
       return {
